requirements.txt
alembic==0.8.4
appnope==0.1.0
bcrypt==2.0.0
blinker==1.4
bottlenose==0.6.3
cffi==1.5.0
click==6.2
decorator==4.0.6
dominate==2.1.16
Flask==0.10.1
flask-appconfig==0.11.1
Flask-Bootstrap==3.3.5.7
Flask-Login==0.3.2
Flask-Mail==0.9.1
Flask-Migrate==1.7.0
flask-nav==0.5
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
Flask-User==0.6.8
Flask-WTF==0.12
gnureadline==6.3.3
gunicorn==19.4.5
ipython==4.0.3
ipython-genutils==0.1.0
itsdangerous==0.24
Jinja2==2.8
lxml==3.5.0
Mako==1.0.3
MarkupSafe==0.23
passlib==1.6.5
path.py==8.1.2
pexpect==4.0.1
pickleshare==0.6
psycopg2==2.6.1
ptyprocess==0.5
pycparser==2.14
pycrypto==2.6.1
python-amazon-simple-product-api==2.0.1
python-dateutil==2.4.2
python-editor==0.5
simplegeneric==0.8.1
six==1.10.0
SQLAlchemy==1.0.11
traitlets==4.1.0
visitor==0.1.2
Werkzeug==0.11.3
wheel==0.24.0
WTForms==2.1
